Plugged Credits Screen in; Refactored;
