--- client/src/TitleScreen/Components/Credits.html
@@ -1,18 +1,0 @@
-<div class="credits-container slide-right">
-    <div class="icons">
-        <img src="../Assets/Icons/close.svg" alt="" />
-    </div>
-    <div class="Credits-Heading">Credits</div>
-        <div class="credit-list">
-            <p class="credit-title">Lead Developer</p>
-            <p class="credit-name">Faraz Sheikh</p>
-            <p class="credit-title">Game Developer</p>
-            <p class="credit-name">Kishan Bapodra</p>
-            <p class="credit-title">Back-end Developer</p>
-            <p class="credit-name">Mostafa Elsayed</p>
-            <p class="credit-title">Front-end Developer</p>
-            <p class="credit-name">Gaurav Nayar</p>
-        </div>
-</div>
-    
-
