+++ client/src/Views/TitleScreen/Components/Controls.html
@@ -1,6 +1,6 @@
 <div class="controls-container slide-right">
 	<div class="icons">
-		<img src="../Assets/Icons/close.svg" alt="" />
+		<img src="../../Assets/Icons/close.svg" alt="" />
 	</div>
 	<div class="controls">
 		<table>
@@ -18,39 +18,42 @@
 				<tr>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/Keyboard/Space-Key.webm" type="video/webm" />
+							<source
+								src="../../Assets/gifs/Keyboard/Space-Key.webm"
+								type="video/webm"
+							/>
 						</video>
 					</td>
 					<td><div class="text crt">&emsp;Boost&emsp;</div></td>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/PS4/X.webm" type="video/webm" />
+							<source src="../../Assets/gifs/PS4/X.webm" type="video/webm" />
 						</video>
 					</td>
 				</tr>
 				<tr>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/Keyboard/A-Key.webm" type="video/webm" />
+							<source src="../../Assets/gifs/Keyboard/A-Key.webm" type="video/webm" />
 						</video>
 					</td>
 					<td><div class="text crt">&emsp;Rotate Left&emsp;</div></td>
 					<td>
-						<video autoplay muted playsinline loop>
-							<source src="../Assets/gifs/PS4/L1.webm" type="video/webm" />
+						<video autoplay muted playsinline loop class="scaleDown">
+							<source src="../../Assets/gifs/PS4/L1.webm" type="video/webm" />
 						</video>
 					</td>
 				</tr>
 				<tr>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/Keyboard/D-Key.webm" type="video/webm" />
+							<source src="../../Assets/gifs/Keyboard/D-Key.webm" type="video/webm" />
 						</video>
 					</td>
 					<td><div class="text crt">&emsp;Rotate Right&emsp;</div></td>
 					<td>
-						<video autoplay muted playsinline loop>
-							<source src="../Assets/gifs/PS4/R1.webm" type="video/webm" />
+						<video autoplay muted playsinline loop class="scaleDown">
+							<source src="../../Assets/gifs/PS4/R1.webm" type="video/webm" />
 						</video>
 					</td>
 				</tr>
@@ -67,26 +70,26 @@
 				<tr>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/Keyboard/E-Key.webm" type="video/webm" />
+							<source src="../../Assets/gifs/Keyboard/E-Key.webm" type="video/webm" />
 						</video>
 					</td>
 					<td><div class="text crt">&emsp;Info&emsp;</div></td>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/PS4/TRIANGLE.webm" type="video/webm" />
+							<source src="../../Assets/gifs/PS4/TRIANGLE.webm" type="video/webm" />
 						</video>
 					</td>
 				</tr>
 				<tr>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/Keyboard/F-Key.webm" type="video/webm" />
+							<source src="../../Assets/gifs/Keyboard/F-Key.webm" type="video/webm" />
 						</video>
 					</td>
 					<td><div class="text crt">&emsp;Shoot&emsp;</div></td>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/PS4/CIRCLE.webm" type="video/webm" />
+							<source src="../../Assets/gifs/PS4/CIRCLE.webm" type="video/webm" />
 						</video>
 					</td>
 				</tr>
@@ -103,26 +106,26 @@
 				<tr>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/Keyboard/Z-Key.webm" type="video/webm" />
+							<source src="../../Assets/gifs/Keyboard/Z-Key.webm" type="video/webm" />
 						</video>
 					</td>
 					<td><div class="text crt">&emsp;Rewind Track&emsp;</div></td>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/PS4/LEFT.webm" type="video/webm" />
+							<source src="../../Assets/gifs/PS4/LEFT.webm" type="video/webm" />
 						</video>
 					</td>
 				</tr>
 				<tr>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/Keyboard/C-Key.webm" type="video/webm" />
+							<source src="../../Assets/gifs/Keyboard/C-Key.webm" type="video/webm" />
 						</video>
 					</td>
 					<td><div class="text crt">&emsp;Next Track&emsp;</div></td>
 					<td>
 						<video autoplay muted playsinline loop class="scaleDown">
-							<source src="../Assets/gifs/PS4/RIGHT.webm" type="video/webm" />
+							<source src="../../Assets/gifs/PS4/RIGHT.webm" type="video/webm" />
 						</video>
 					</td>
 				</tr>
