+++ client/src/Views/TitleScreen/Components/Credits.html
@@ -1,0 +1,24 @@
+<div class="credits-container slide-right">
+	<div class="icons">
+		<img src="../../Assets/Icons/close.svg" alt="" />
+	</div>
+	<div class="Credits-Heading crt">Credits</div>
+	<div class="credit-list">
+		<div class="heading crt">Developers</div>
+
+		<table>
+			<tbody>
+				<tr>
+					<td class="title crt">Lead Developer</td>
+					<td class="heading crt">:</td>
+					<td class="text crt">Faraz Shaikh</td>
+				</tr>
+				<tr>
+					<td class="title crt">Gameplay Developer</td>
+					<td class="heading crt">:</td>
+					<td class="text crt">Kishan Bapodra</td>
+				</tr>
+			</tbody>
+		</table>
+	</div>
+</div>
