+++ client/src/Views/TitleScreen/Components/Play.html
@@ -1,6 +1,6 @@
 <div class="play-container slide-right">
 	<div class="icons">
-		<img src="../Assets/Icons/close.svg" alt="" />
+		<img src="../../Assets/Icons/close.svg" alt="" />
 	</div>
 	<div class="title crt">Registration</div>
 	<div class="play-container-holder">
