+++ client/src/Views/TitleScreen/TitleScreen.html
@@ -22,7 +22,7 @@
 		</div>
 
 		<div class="icons">
-			<img src="./Assets/Icons/Mute.svg" alt="" />
+			<img src="../Assets/Icons/Mute.svg" alt="" />
 		</div>
 	</div>
 	<div class="inject-html"></div>
